moved some constants from src/mock/const.js to src/const.js
