+++ src/const.js
@@ -1,1 +1,119 @@
 export const POINT_TYPES = ['taxi', 'bus', 'train', 'ship', 'drive', 'flight', 'check-in', 'sightseeing', 'restaurant'];
+export const OFFER_TYPES = [
+  {
+    'type': 'taxi',
+    'offers': [
+      {
+        'title': 'Order Uber',
+        'price': 20,
+      },
+      {
+        'title': 'Upgrade to a business class',
+        'price': 20,
+      },
+      {
+        'title': 'Choose the radio station',
+        'price': 10,
+      },
+    ],
+  },
+  {
+    'type': 'bus',
+    'offers': [
+      {
+        'title': 'Choose meal',
+        'price': 180,
+      },
+      {
+        'title': 'Upgrade to comfort class',
+        'price': 50,
+      },
+    ],
+  },
+  {
+    'type': 'train',
+    'offers': [],
+  },
+  {
+    'type': 'ship',
+    'offers': [
+      {
+        'title': 'Add meal',
+        'price': 15,
+      },
+      {
+        'title': 'Switch to comfort class',
+        'price': 100,
+      },
+      {
+        'title': 'Add luggage',
+        'price': 30,
+      },
+    ],
+  },
+  {
+    'type': 'drive',
+    'offers': [
+      {
+        'title': 'Rent a car',
+        'price': 200,
+      },
+    ],
+  },
+  {
+    'type': 'flight',
+    'offers': [
+      {
+        'title': 'Add meal',
+        'price': 15,
+      },
+      {
+        'title': 'Switch to comfort class',
+        'price': 100,
+      },
+      {
+        'title': 'Choose seats',
+        'price': 5,
+      },
+      {
+        'title': 'Add luggage',
+        'price': 30,
+      },
+      {
+        'title': 'Travel by train',
+        'price': 40,
+      },
+    ],
+  },
+  {
+    'type': 'check-in',
+    'offers': [
+      {
+        'title': 'Add breakfast',
+        'price': 50,
+      },
+      {
+        'title': 'Add dinner',
+        'price': 50,
+      },
+    ],
+  },
+  {
+    'type': 'sightseeing',
+    'offers': [
+      {
+        'title': 'Book tickets',
+        'price': 40,
+      },
+      {
+        'title': 'Lunch in city',
+        'price': 30,
+      },
+    ],
+  },
+  {
+    'type': 'restaurant',
+    'offers': [],
+  },
+];
+export const CITIES = ['Amsterdam', 'Chamonix', 'Geneva', 'Prague', 'Berlin', 'Stockholm'];
