+++ src/mock/const.js
@@ -12,124 +12,6 @@ export const DESCRIPTIONS = [
   'In rutrum ac purus sit amet tempus.',
 ];
 
-export const OFFER_TYPES = [
-  {
-    'type': 'taxi',
-    'offers': [
-      {
-        'title': 'Order Uber',
-        'price': 20,
-      },
-      {
-        'title': 'Upgrade to a business class',
-        'price': 20,
-      },
-      {
-        'title': 'Choose the radio station',
-        'price': 10,
-      },
-    ],
-  },
-  {
-    'type': 'bus',
-    'offers': [
-      {
-        'title': 'Choose meal',
-        'price': 180,
-      },
-      {
-        'title': 'Upgrade to comfort class',
-        'price': 50,
-      },
-    ],
-  },
-  {
-    'type': 'train',
-    'offers': [],
-  },
-  {
-    'type': 'ship',
-    'offers': [
-      {
-        'title': 'Add meal',
-        'price': 15,
-      },
-      {
-        'title': 'Switch to comfort class',
-        'price': 100,
-      },
-      {
-        'title': 'Add luggage',
-        'price': 30,
-      },
-    ],
-  },
-  {
-    'type': 'drive',
-    'offers': [
-      {
-        'title': 'Rent a car',
-        'price': 200,
-      },
-    ],
-  },
-  {
-    'type': 'flight',
-    'offers': [
-      {
-        'title': 'Add meal',
-        'price': 15,
-      },
-      {
-        'title': 'Switch to comfort class',
-        'price': 100,
-      },
-      {
-        'title': 'Choose seats',
-        'price': 5,
-      },
-      {
-        'title': 'Add luggage',
-        'price': 30,
-      },
-      {
-        'title': 'Travel by train',
-        'price': 40,
-      },
-    ],
-  },
-  {
-    'type': 'check-in',
-    'offers': [
-      {
-        'title': 'Add breakfast',
-        'price': 50,
-      },
-      {
-        'title': 'Add dinner',
-        'price': 50,
-      },
-    ],
-  },
-  {
-    'type': 'sightseeing',
-    'offers': [
-      {
-        'title': 'Book tickets',
-        'price': 40,
-      },
-      {
-        'title': 'Lunch in city',
-        'price': 30,
-      },
-    ],
-  },
-  {
-    'type': 'restaurant',
-    'offers': [],
-  },
-];
-export const POINT_TYPES = ['taxi', 'bus', 'train', 'ship', 'drive', 'flight', 'check-in', 'sightseeing', 'restaurant'];
 export const POINT_TYPE_BASE_PRICE = {
   'taxi': 200,
   'bus': 1100,
@@ -142,4 +24,3 @@ export const POINT_TYPE_BASE_PRICE = {
   'restaurant': 80,
 };
 export const PHOTOS_URL = 'http://picsum.photos/248/152?r=';
-export const CITIES = ['Amsterdam', 'Chamonix', 'Geneva', 'Prague', 'Berlin', 'Stockholm'];
