+++ src/view/edit-event-form.js
@@ -1,6 +1,6 @@
 import AbstractView from './abstract';
 import {POINT_TYPES} from '../const';
-import {CITIES, OFFER_TYPES} from '../mock/const';
+import {CITIES, OFFER_TYPES} from '../const';
 import {getHumanizeVisibleDateForForm} from '../utils/date-format';
